remove: MyPageView.tsx삭제 후 해당 기능들을 MyPage.tsx로 이동 Github issue #163
